+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Bug Fixes
 - Allow users with only annotation access to edit annotations ([#1267](../../pull/1267))
+- Fix filtering item lists in Girder by multiple phrases ([#1272](../../pull/1272))
 
 ## 1.23.3